move load_file to helpers
